pyproject.toml
[project]
name = "mapcli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml >=6.0",
    "copier >=9,<10"
]

[project.scripts]
mapcli = "mapcli.create_cli_pkg:main"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
